Calculate correlations between a dependent and multiple independent variables

Calculate correlations between a dependent and multiple independent variables

ryx(data, y, x)

Arguments

data

A data frame containing the variables.

y

The dependent variable (character).

x

The independent variables (character vector).

Value

An object of class `ryx` with results.

Examples

library(MASS)
x <- ryx(Boston, y="medv")
print(x)
#> Correlations of medv with
#>  variable          r            p sigif
#>     lstat -0.7376627 5.081103e-88   ***
#>        rm  0.6953599 2.487229e-74   ***
#>   ptratio -0.5077867 1.609509e-34   ***
#>     indus -0.4837252 4.900260e-31   ***
#>       tax -0.4685359 5.637734e-29   ***
#>       nox -0.4273208 7.065042e-24   ***
#>      crim -0.3883046 1.173987e-19   ***
#>       age -0.3769546 1.569982e-18   ***
#>        zn  0.3604453 5.713584e-17   ***
#>     black  0.3334608 1.318113e-14   ***
#>       dis  0.2499287 1.206612e-08   ***